Migration towards Java version of 14/06/2019 - Moved core sub-packages into 'core'.

+++ java/component/keyple-common/src/main/java/org/eclipse/keyple/common/System.h
@@ -33,6 +33,6 @@ public:
 
     static long currentTimeMillis()
 	{
-        return nanoTime() * pow(10, 6);
+        return (long)(nanoTime() * pow(10, 6));
     }
 };
